split test_update_images file
